aider: Moved the search box to the end of the navigation bar.
# Aider chat conversation:

USER: The search box is under the nav bar when it should be at the end of the nav bar

ASSISTANT: To move the search box to the end of the navigation bar, we need to modify the `src/components/NavigationBar.js` and `src/App.css` files.

Here are the *SEARCH/REPLACE* blocks:

src/components/NavigationBar.js
```javascript
<<<<<<< SEARCH
return (
    <nav>
        <ul className="navbar">
            <li><Link to="/">Home</Link></li>
            <li><Link to="/about">About</Link></li>
            <li><Link to="/contact">Contact</Link></li>
        </ul>
        <SearchBar />
    </nav>
);
=======
return (
    <nav className="navbar">
        <ul>
            <li><Link to="/">Home</Link></li>
            <li><Link to="/about">About</Link></li>
            <li><Link to="/contact">Contact</Link></li>
        </ul>
        <SearchBar />
    </nav>
);
>>>>>>> REPLACE
```

src/App.css
```css
<<<<<<< SEARCH
.navbar {
    list-style-type: none;
    margin: 0;
    padding: 0;
    overflow: hidden;
    background-color: #333;
    display: flex;
    justify-content: space-between;
    align-items: center;
}
=======
.navbar {
    list-style-type: none;
    margin: 0;
    padding: 0;
    overflow: hidden;
    background-color: #333;
    display: flex;
    justify-content: flex-start;
    align-items: center;
}

.navbar ul {
    display: flex;
}
>>>>>>> REPLACE
```

In the `NavigationBar.js` file, we moved the `navbar` class from the `ul` element to the `nav` element. Then, in the `App.css` file, we changed the `justify-content` property of the `navbar` class from `space-between` to `flex-start` to align the items to the start of the container. We also added a `display: flex;` rule to the `navbar ul` selector to make the list items display in a row.