aider: Formatted the logo to be smaller by adding a new CSS class and applying it to the logo.
# Aider chat conversation:

USER: The logo is massive and takes up the whole page you need to format it to be small

ASSISTANT: To make the logo smaller, we need to modify the `src/App.css` file to add a new CSS class for the logo and then apply this class to the logo in the `src/components/NavigationBar.js` file.

Here are the *SEARCH/REPLACE* blocks:

src/App.css
```css
<<<<<<< SEARCH
.navbar li a:hover {
    background-color: #1db954;
    color: #b3b3b3;
}
=======
.navbar li a:hover {
    background-color: #1db954;
    color: #b3b3b3;
}

.navbar-logo {
    height: 50px;
    width: 50px;
}
>>>>>>> REPLACE
```

src/components/NavigationBar.js
```javascript
<<<<<<< SEARCH
<img src={logo} alt="Logo" />
=======
<img src={logo} alt="Logo" className="navbar-logo" />
>>>>>>> REPLACE
```

This will make the logo 50px by 50px. You can adjust the height and width values in the `.navbar-logo` class in `src/App.css` to make the logo the size you want.