aider: Make the white background of the logo transparent using the `mix-blend-mode` property.
# Aider chat conversation:

USER: The background of the logo is white, can you edit that?

ASSISTANT: As an AI, I'm unable to directly edit image files. However, I can suggest a few ways you can remove the white background from your logo:

1. **Use an image editing software:** You can use software like Adobe Photoshop, GIMP, or online tools like remove.bg to remove the background from your image.

2. **Use a version of your logo with a transparent background:** If you have a version of your logo with a transparent background (usually a .png file), you can replace the current logo file with that.

3. **CSS Hack:** If the white background is a solid color, you can use a CSS hack to make it transparent. This won't work perfectly, especially if there are other white parts in the image that you don't want to be transparent.

Once you've edited your logo, you can replace the current logo file in your project with the new one. The path to replace the logo file would be `src/assets/circle_leaf_logo.jpg`.

USER: Use a css hack to make it transparent

ASSISTANT: To make the white background of the logo transparent using CSS, we can use the `mix-blend-mode` property. This property describes how an element's content should blend with the content of the element's parent and the element's background.

Here is the *SEARCH/REPLACE* block:

src/App.css
```css
<<<<<<< SEARCH
.navbar-logo {
    height: 50px;
    width: 50px;
}
=======
.navbar-logo {
    height: 50px;
    width: 50px;
    mix-blend-mode: multiply;
}
>>>>>>> REPLACE
```

This will make the white background of the logo transparent. Please note that this is a hack and may not work perfectly in all scenarios. It's always better to use an image with a transparent background for best results.