aider: Add route for Spotify page in `src/App.js`.
# Aider chat conversation:

USER: I do not see anything on the spotify page

ASSISTANT: It seems like the Spotify page is not included in the routing of your application. To fix this, we need to add a route for the Spotify page in the `src/App.js` file.

Here is the *SEARCH/REPLACE* block:

src/App.js
```python
<<<<<<< SEARCH
        <Route path="/register" element={<RegisterPage />} />
=======
        <Route path="/register" element={<RegisterPage />} />
        <Route path="/spotify" element={<Spotify />} />
>>>>>>> REPLACE
```

This change will add a new route `/spotify` that will render the `Spotify` component when navigated to. Please make sure that the `Spotify` component is imported at the top of the `src/App.js` file. If it's not, you will need to add `import Spotify from './components/Spotify';` to the list of imports.