aider: Added import statement for the Spotify component in the App.js file.
# Aider chat conversation:

USER: ERROR
[eslint]
src/App.js
  Line 23:42:  'Spotify' is not defined  react/jsx-no-undef

Search for the keywords to learn more about each error.

ASSISTANT: The error message indicates that the `Spotify` component is not defined in the `src/App.js` file. This is because the `Spotify` component is not imported in this file.

To fix this, we need to add an import statement for the `Spotify` component at the top of the `src/App.js` file.

Here is the *SEARCH/REPLACE* block:

src/App.js
```python
<<<<<<< SEARCH
import RegisterPage from './pages/RegisterPage';
=======
import RegisterPage from './pages/RegisterPage';
import Spotify from './components/Spotify';
>>>>>>> REPLACE
```

This change will import the `Spotify` component from the `src/components/Spotify.js` file.