aider: Changed the database name in `server/db.js` to `Tunify` to fix the database connection issue.
# Aider chat conversation:

USER: What is /sampleData?

ASSISTANT: The `/sampleData` is a route in your Express.js server defined in `server/routes/spotifyRoutes.js`. When a GET request is made to this route, it triggers the `getSampleData` function from `server/controllers/spotifyController.js`.

This function executes a SQL query to fetch all records from the `Tunify` table in your database. The results are then sent back as a response to the client who made the request.

In your React component `src/components/Spotify.js`, an API call is made to this `/sampleData` route when the component is first rendered. The data received from this API call is then stored in the component's state and used to display information on the page.

USER: Follow the route and controller to find why database connection isnt being made correctly

ASSISTANT: To troubleshoot the database connection issue, we need to look at the `server/controllers/spotifyController.js` and `server/routes/spotifyRoutes.js` files. However, the actual database connection is usually established in a separate file, often named `db.js` or similar. This file is not currently listed as a *read-write* file.

Could you please add the database connection file to the chat so we can examine it? It's usually where the connection to the database is established and exported for use in other parts of the application, like the `spotifyController.js` file.

USER: I added these *read-write* files: server/db.js

ASSISTANT: Thank you for providing the `server/db.js` file.

Looking at the `server/db.js` file, it seems that the database connection is being established with the following details:

```javascript
const db = mysql.createConnection({
    host: 'localhost',
    user: 'root',
    password: 'password',
    database: 'myDatabase'
});
```

However, in your `server/controllers/spotifyController.js` file, you're trying to fetch data from a database named `Tunify`:

```javascript
db.query('SELECT * FROM Tunify', (err, results) => { ... });
```

This discrepancy could be the reason why the database connection isn't being made correctly.

To fix this, you should change the database name in `server/db.js` to `Tunify` if that's the correct database you want to connect to.

Here is the *SEARCH/REPLACE* block:

server/db.js
```javascript
<<<<<<< SEARCH
    database: 'myDatabase'
=======
    database: 'Tunify'
>>>>>>> REPLACE
```

Please make sure that the database `Tunify` exists in your MySQL server and the `root` user has the correct privileges to access it. Also, ensure that the password for the `root` user is correct.